// src/server/v1/classifieds/dtos/create-images.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsEnum, IsNotEmpty, IsString } from 'class-validator';

export class CreateImagesDto {
    @ApiProperty({
        description: 'Etag',
        example: 'random etag string auto generated',
    })
    @IsNotEmpty()
    @IsString()
    etag: string;

    @ApiProperty({
        description: 'url',
        example: 'url generated by contabo',
    })
    @IsNotEmpty()
    @IsString()
    url: string;

    @ApiProperty({
        description: 'file',
        example: 'example file url',
    })
    @IsNotEmpty()
    @IsString()
    file: string;
}